requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
gevent==24.2.1
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
pycryptodome==3.20.0
python-dotenv==1.0.1
rfc3986==1.5.0
sniffio==1.3.1
Werkzeug==3.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
beautifulsoup4==4.12.3
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.7.4
cffi==1.16.0
cssselect2==0.7.0
curl_cffi==0.7.1
defusedxml==0.7.1
g4f==0.3.2.4
idna==2.10
pillow==10.4.0
pycparser==2.22
requests==2.32.3
soupsieve==2.5
tinycss2==1.3.0
urllib3==2.2.2
webencodings==0.5.1